requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.0
openai==1.12.0
supabase==2.0.0
python-jose==3.3.0
httpx==0.24.0
postgrest==0.10.8
pydantic==2.1.1
pydantic-settings==2.0.3
